restructurize response utils
