+++ src/commonMain/kotlin/eu/vendeli/rethis/commands/ScriptingCommands.kt
@@ -5,8 +5,8 @@ import eu.vendeli.rethis.types.core.RType
 import eu.vendeli.rethis.types.core.toArg
 import eu.vendeli.rethis.types.core.unwrapList
 import eu.vendeli.rethis.types.options.FunctionRestoreOption
+import eu.vendeli.rethis.utils.response.unwrapRespIndMap
 import eu.vendeli.rethis.utils.safeCast
-import eu.vendeli.rethis.utils.unwrapRespIndMap
 import eu.vendeli.rethis.utils.writeArg
 
 suspend fun ReThis.eval(script: String, numKeys: Long, vararg keys: String): RType = execute(
