+++ src/commonMain/kotlin/eu/vendeli/rethis/commands/StreamCommands.kt
@@ -6,7 +6,7 @@ import eu.vendeli.rethis.types.core.RType
 import eu.vendeli.rethis.types.core.toArg
 import eu.vendeli.rethis.types.core.unwrapList
 import eu.vendeli.rethis.types.options.*
-import eu.vendeli.rethis.utils.unwrapRespIndMap
+import eu.vendeli.rethis.utils.response.unwrapRespIndMap
 import eu.vendeli.rethis.utils.writeArg
 
 suspend fun ReThis.xAck(
